add controllers

+++ src/main/java/com/ada/ecommerce/aulatdd/model/Venda.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -18,6 +19,6 @@ public class Venda {
     @Id
     private Long id;
     private String historico;
-    private Double saldo;
+    private BigDecimal saldo;
     private LocalDateTime dataVenda;
 }
